adding in "scaled + centered" data calculation
Originally, the calculation of valscmat had to be rerun (originally Line 176) depending on if you wanted to have the data centered + scaled (underlying Figure 6) or just scaled (for SPAC).

I added valzmat to handle the "centered + scaled" data so the the code only has to be run once and adjusted the lines in Step 5 (Volatility) to use the centered data.